Adding code for floating point multiply
This code is adapated from
https://opencores.org/websvn/listing?repname=fpuvhdl&path=%2Ffpuvhdl%2F&rev=0

This has not been thoroughly tested and does not even pass the current
testbench. It currently has issues with how rounding is handled. However
some welcome changes are the number of bits dedicated to the significand
(mantissa) and the exponent are parameterized, giving us fine-grained
control over resource usage.

It would be a good idea to a) add addition for the accumulator b)
integrate the operators into the actual matvec design c) create a
testbench that tests all of the floating point standard (i.e.
subnormals, rounding, zero, etc.).